requirements.txt
django
djangorestframework
drf-spectacular
pillow
drf_yasg